requirements.txt
affine==2.4.0
anyio==3.6.2
attrs==23.1.0
blinker==1.6.2
boto3==1.26.137
botocore==1.29.137
cachetools==5.3.0
certifi==2023.5.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.1
Flask==2.3.2
GDAL==3.6.4
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
morecantile==3.4.0
numexpr==2.8.4
numpy==1.24.3
Pillow==9.5.0
pydantic==1.10.7
pyparsing==3.0.9
pyproj==3.5.0
pystac==1.7.3
python-dateutil==2.8.2
rasterio==1.3.6
rio-tiler==4.1.11
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
zipp==3.15.0
